fix: restore documentation baseline and verification shim
