+++ app/content/japanese_quality.py
