+++ app/content/script_quality.py
@@ -13,7 +13,7 @@ from typing import Any, Dict, List
 
 import google.generativeai as genai
 
-from .config import cfg
+from app.config import cfg
 
 logger = logging.getLogger(__name__)
