+++ app/core/__init__.py
@@ -1,0 +1,20 @@
+"""Core orchestration and shared infrastructure modules."""
+
+from .api_rotation import APIKey, APIKeyRotationManager, get_rotation_manager, initialize_api_infrastructure
+from .container import AppContainer
+from .logging_config import WorkflowLogger, get_log_session, setup_logging
+from .utils import DateUtils, FileUtils, TextUtils
+
+__all__ = [
+    "APIKey",
+    "APIKeyRotationManager",
+    "get_rotation_manager",
+    "initialize_api_infrastructure",
+    "AppContainer",
+    "WorkflowLogger",
+    "get_log_session",
+    "setup_logging",
+    "DateUtils",
+    "FileUtils",
+    "TextUtils",
+]
