+++ app/integrations/__init__.py
@@ -1,0 +1,20 @@
+"""External service integrations used throughout the workflow."""
+
+from .discord import DiscordNotifier, discord_notifier
+from .drive import DriveManager as DriveUploader
+from .metadata import MetadataGenerator
+from .metadata_storage import metadata_storage
+from .sheets import sheets_manager
+from .web import create_app
+from .youtube import YouTubeManager
+
+__all__ = [
+    "DiscordNotifier",
+    "discord_notifier",
+    "DriveUploader",
+    "MetadataGenerator",
+    "metadata_storage",
+    "sheets_manager",
+    "create_app",
+    "YouTubeManager",
+]
