+++ app/news/__init__.py
@@ -1,0 +1,5 @@
+"""News collection and enrichment services."""
+
+from .search import NewsCollector, collect_news
+
+__all__ = ["NewsCollector", "collect_news"]
