+++ tests/integration/test_youtube_uploads.py
@@ -12,7 +12,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__))))
 import logging
 
 from app.config import cfg
-from app.youtube import YouTubeManager
+from app.integrations.youtube import YouTubeManager
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
 logger = logging.getLogger(__name__)
