+++ tests/unit/test_script_parse.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
-from app.tts import tts_manager
+from app.media.tts import tts_manager
 
 # Test with proper script format (what CrewAI should generate)
 test_script = """
